:gear: refactor cli commands
