+++ cli/src/cmd/cast/exec.rs
@@ -1,4 +1,4 @@
-use crate::cmd::{build::BuildArgs, Cmd};
+use crate::cmd::{forge::build::BuildArgs, Cmd};
 use clap::Parser;
 
 use forge::ContractRunner;
