+++ cli/src/forge.rs
@@ -3,7 +3,10 @@ mod opts;
 mod term;
 mod utils;
 
-use crate::cmd::{watch, Cmd};
+use crate::cmd::{
+    forge::{verify, watch},
+    Cmd,
+};
 
 use ethers::solc::{Project, ProjectPathsConfig};
 use opts::forge::{Dependency, Opts, Subcommands};
@@ -31,7 +34,7 @@ fn main() -> eyre::Result<()> {
         }
         Subcommands::Build(cmd) => {
             if cmd.is_watch() {
-                utils::block_on(crate::cmd::watch::watch_build(cmd))?;
+                utils::block_on(watch::watch_build(cmd))?;
             } else {
                 cmd.run()?;
             }
@@ -39,14 +42,14 @@ fn main() -> eyre::Result<()> {
         Subcommands::Run(cmd) => {
             cmd.run()?;
         }
-        Subcommands::Exec(cmd) => {
-            cmd.run()?;
-        }
+        // Subcommands::Exec(cmd) => {
+        //     cmd.run()?;
+        // }
         Subcommands::VerifyContract(args) => {
-            utils::block_on(cmd::verify::run_verify(&args))?;
+            utils::block_on(verify::run_verify(&args))?;
         }
         Subcommands::VerifyCheck(args) => {
-            utils::block_on(cmd::verify::run_verify_check(&args))?;
+            utils::block_on(verify::run_verify_check(&args))?;
         }
         Subcommands::Create(cmd) => {
             cmd.run()?;
